=act move lookupAndSetField to Reflect helper.
